Add submenu for 16 / 20Mhz
This change allows to run Nano Every at full speed, i.e. 20Mhz. It sets the right FUSE flags and CPU frequency macros - as already suggested by several people (for example  https://forum.arduino.cc/t/clock-speed-for-nano-every/599449/7).
If I understand correctly, this may or may not work with people's projects - any way we should allow people to give it a try. After all, one strength of the arduino platform is that it's "open for hacking' ;-).